Automated ARDUINO_LIBS collection from Ed Comer.
Ed writes:

I made another change. I got tired of setting up a make process and discovering
that  ARDUINO_LIBS was wrong, so I automated that also. Now the makefile
searches the source file for includes, sees if they are resident in the
libraries folder and if they are, creates a ARDUINO_LIBS variable populated with
the libraries imported by the source file. I've attached my current Arduino.mk